chore: move block style to correct place
