+++ packages/marketplace/src/components/pages/app-detail/__styles__/standalone-app-detail.scss
@@ -1,6 +1,6 @@
-@import '../base/colors.scss';
-@import '../base/screen-size.scss';
-@import '../base/browser.scss';
+@import '@/styles/base/colors.scss';
+@import '@/styles/base/screen-size.scss';
+@import '@/styles/base/browser.scss';
 
 .preview {
   position: relative;
