+++ packages/marketplace/src/components/pages/app-detail/client/app-detail.tsx
@@ -12,7 +12,7 @@ import { selectIsAdmin } from '@/selector/auth'
 import { canGoBack } from '@/utils/router-helper'
 import AppContent from './app-content'
 import { Loader, GridItem, Grid, Section } from '@reapit/elements'
-import styles from '@/styles/blocks/standalone-app-detail.scss?mod'
+import styles from '../__styles__/standalone-app-detail.scss?mod'
 import AppInstallConfirmation from '@/components/pages/app-detail/client/app-install-confirmation'
 import { Aside } from './aside'
 import { getDesktopIntegrationTypes } from '@/utils/get-desktop-integration-types'
