+++ packages/marketplace/src/components/pages/app-detail/common/ui-app-header.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { H3, Grid, GridItem, SubTitleH6 } from '@reapit/elements'
 import { FaCheck } from 'react-icons/fa'
 import { AppDetailModel } from '@reapit/foundations-ts-definitions'
-import styles from '@/styles/blocks/standalone-app-detail.scss?mod'
+import styles from '../__styles__/standalone-app-detail.scss?mod'
 import { MEDIA_INDEX } from '@/constants/media'
 import ImagePlaceHolder from '@/assets/images/default-app-icon.jpg'
 import featureImagePlaceHolder from '@/assets/images/default-feature-image.jpg'
