+++ packages/marketplace/src/components/pages/installed-apps/__styles__/installed-app-card.scss
