+++ packages/marketplace/src/components/ui/app-card/__styles__/app-card.scss
@@ -1,4 +1,4 @@
-@import '../base/colors.scss';
+@import '@/styles/base/colors.scss';
 
 // This solution is a bit shit and will only add an elipsis in webkit browsers
 // Sorry but other browsers no-supporty...
