+++ packages/marketplace/src/components/ui/app-list/__styles__/app-list.scss
