+++ packages/marketplace/src/components/ui/app-list/index.ts
@@ -1,0 +1,1 @@
+export { default } from './app-list'
