+++ packages/marketplace/src/components/ui/app-sidebar/app-sidebar.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { useSelector } from 'react-redux'
 import { FlexContainerBasic, H5, Input, Formik, Form, RadioSelect, FormikProps, H6 } from '@reapit/elements'
 import { useHistory, useLocation } from 'react-router'
-import styles from '@/styles/blocks/app-sidebar.scss?mod'
+import styles from './__styles__/app-sidebar.scss?mod'
 import CategoriesList from '@/components/ui/categories-list'
 import { FaSearch } from 'react-icons/fa'
 import { CategoryModel } from '@reapit/foundations-ts-definitions'
