+++ packages/marketplace/src/components/ui/categories-list/__styles__/categories-list.scss
@@ -1,0 +1,6 @@
+.listCategories {
+  @include for-mobile-only {
+    display: flex;
+    flex-wrap: wrap;
+  }
+}
