+++ packages/marketplace/src/components/ui/categories-list/__tests__/categories-list.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
 
-import CategoriesList, { CategoriesListProps } from '@/components/ui/categories-list'
+import CategoriesList, { CategoriesListProps } from '../categories-list'
 import CategoryItem from '@/components/ui/category-item'
 import { categoriesStub } from '@/sagas/__stubs__/categories'
 import { CategoryModel } from '@reapit/foundations-ts-definitions'
