+++ packages/marketplace/src/components/ui/categories-list/categories-list.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
-import CategoryItem from './category-item'
-import styles from '@/styles/blocks/categories-list.scss?mod'
+import CategoryItem from '../category-item'
+import styles from './__styles__/categories-list.scss?mod'
 import { CategoryModel } from '@reapit/foundations-ts-definitions'
 
 export interface CategoriesListProps {
