+++ packages/marketplace/src/components/ui/categories-list/index.ts
@@ -1,0 +1,1 @@
+export { default } from './categories-list'
