+++ packages/marketplace/src/components/ui/category-item/category-item.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import styles from '@/styles/blocks/categories-list.scss?mod'
+import styles from './__styles__/categories-item.scss?mod'
 import { CategoryModel } from '@reapit/foundations-ts-definitions'
 
 export interface CategoryItemProps {
