+++ packages/marketplace/src/components/ui/category-item/index.ts
@@ -1,0 +1,1 @@
+export { default } from './category-item'
