+++ packages/marketplace/src/components/ui/featured-app/__styles__/featured-app.scss
