+++ packages/marketplace/src/components/ui/featured-app/__tests__/featured-app.tsx
