+++ packages/marketplace/src/components/ui/featured-app/index.ts
@@ -1,0 +1,1 @@
+export { default } from './featured-app'
