--- packages/marketplace/src/styles/blocks/app-authentication-detail.scss
@@ -1,58 +1,0 @@
-@import '../base/colors.scss';
-
-.btnShowAuthentication {
-  text-decoration: underline;
-}
-
-.authenticationCodeWrap {
-  padding: 8px;
-  border: 1px solid $grey;
-  min-width: 80%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.authenticationCode {
-  word-break: break-all;
-  margin-bottom: 0 !important;
-}
-
-.btnCopy {
-  cursor: pointer;
-  position: relative;
-}
-
-.tooltiptext {
-  top: -5px;
-  bottom: auto;
-  right: 128%;
-
-  visibility: hidden;
-  position: absolute;
-  min-width: 80px;
-  background-color: #555;
-  color: #fff;
-  text-align: center;
-  padding: 5px 0;
-  border-radius: 6px;
-  z-index: 1;
-  opacity: 0;
-  transition: opacity 0.6s;
-}
-
-.btnCopy:hover .tooltiptext {
-  visibility: visible;
-  opacity: 1;
-}
-
-.tooltiptext::after {
-  content: '';
-  position: absolute;
-  top: 50%;
-  left: 100%;
-  margin-top: -5px;
-  border-width: 5px;
-  border-style: solid;
-  border-color: transparent transparent transparent #555;
-}
