--- packages/marketplace/src/styles/blocks/loader.scss
@@ -1,51 +1,0 @@
-@import '../base/layout.scss';
-@import '../base/colors.scss';
-
-.loader {
-  display: block;
-  margin: auto;
-  width: $layout-triple;
-  height: $layout-two-half;
-
-  &.body {
-    margin: 100px auto,
-  }
-
-  > div {
-    background-color: $grey;
-    height: 100%;
-    width: $layout-third;
-    margin: 2px;
-    display: inline-block;
-    border-radius: 15%;
-
-    animation: loader-animation 1.2s infinite ease-in-out;
-
-    &:nth-of-type(1) {
-      animation-delay: -1.1s;
-    }
-
-    &:nth-of-type(2) {
-      animation-delay: -1s;
-    }
-
-    &:nth-of-type(3) {
-      animation-delay: -0.9s;
-    }
-
-    &:nth-of-type(4) {
-      animation-delay: -0.8s;
-    }
-  }
-
-  @keyframes loader-animation {
-    0%,
-    40%,
-    100% {
-      transform: scaleY(0.4);
-    }
-    20% {
-      transform: scaleY(1);
-    }
-  }
-}
